+++ Backend/BLL/Sevices/UserService.cs
@@ -12,7 +12,7 @@ public class UserService(
     IMapper mapper,
     PasswordManager passwordManager)
 {
-    public async Task<User?> GetUserByIdAsync(int userId)
+    public async Task<User?> GetUserByIdAsync(long userId)
     {
         var output = await userRepository.GetUserByIdAsync(userId);
         if (output == null) throw new ObjectNotFoundException("User not found. You can't get user that doesn't exist");
@@ -34,7 +34,7 @@ public class UserService(
         return output;
     }
 
-    public async Task<User?> UpdateUserAsync(int id, User userModel)
+    public async Task<User?> UpdateUserAsync(long id, User userModel)
     {
         var user = await userRepository.GetUserByIdAsync(id);
         if (user == null) throw new ObjectNotFoundException("User not found. You can't update user that doesn't exist");
@@ -51,7 +51,7 @@ public class UserService(
         return output;
     }
 
-    public async Task<User?> DeleteUserAsync(int id)
+    public async Task<User?> DeleteUserAsync(long id)
     {
         var user = await userRepository.GetUserByIdAsync(id);
         if (user == null) throw new ObjectNotFoundException("User not found. You can't delete user that doesn't exist");
